+++ ft_printf/handle_uns.c
@@ -6,7 +6,7 @@
 /*   By: hmensah- <hmensah-@student.42abudhabi.a    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/20 15:02:10 by hmensah-          #+#    #+#             */
-/*   Updated: 2025/01/22 17:22:22 by hmensah-         ###   ########.fr       */
+/*   Updated: 2025/01/23 22:36:51 by hmensah-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,10 @@
 
 t_fdata	*handle_uns(t_modinfo *info, va_list args)
 {
-	long	nbr;
-	t_fdata	*data;
+	unsigned int	nbr;
+	t_fdata			*data;
 
-	nbr = va_arg(args, long);
+	nbr = va_arg(args, int);
 	data = create_number_data(info, nbr, 10);
 	apply_precision(info, data);
 	apply_plus_space_flag(info, data);
